+++ o8g/Scripts/constants.py
@@ -1,14 +1,19 @@
           
 mdict = { # A dictionary which holds all the hard coded markers (in the markers file)
-    'exposed'        : ("Exposed marker", "0183c0b8-6a6d-42c5-bedb-94bbc8ca5329"),
-    'baffled'        : ("Baffled marker", "0415bae0-0c88-4ec6-8169-aeb117318618"),
-    'Snoop'          : ("Snoop", "2f9409af-726e-4c57-b65d-1007250c9b81"),
-    'Brief'          : ("Brief", "96aab548-6e02-4509-b583-6cc318469465"),
+    'exposed'        : ("Exposed", "0183c0b8-6a6d-42c5-bedb-94bbc8ca5329"),
+    'baffled'        : ("Baffled", "0415bae0-0c88-4ec6-8169-aeb117318618"),
+    'Snoop'          : ("Used Snoop", "2f9409af-726e-4c57-b65d-1007250c9b81"),
+    'Brief'          : ("Used Brief", "96aab548-6e02-4509-b583-6cc318469465"),
+    'Briefed'        : ("Briefed", "72179c7a-9087-49f8-8a93-010b284cfa6e"),
+    'Fresh'          : ("Fresh Leader", "c964be10-521f-4cc0-a3a9-5c0def322110"),
+    'Demoted'        : ("Demoted", "cdf63a20-808d-4789-b192-da69a870719c"),
+    'DefaultMission' : ("Default Mission Action Taken", "c3dcf14d-65ae-4356-b702-f4b8fda91313"),
     'MissionAction'  : ("Mission Action Taken", "69a2514e-9d49-4f8d-8408-dc069625e07c"),
     'TextAction'     : ("Text Action Taken", "05a5f34f-e4d2-4ee0-9906-b03f70b02a01")
 }
 
-AbilityColor = "#FF0000"
+AttackerColor = "#FF0000"
+DefenderColor = "#0000FF"
 
 CardWidth = 63
 CardHeight = 88