+++ changelog.md
@@ -6,6 +6,7 @@ Changelog - Spycraft CCG OCTGN Game Definition
 * Fixed b0rked mutliplayer mission queue
 * Fixed placement of 2nd player cards and mission queue a bit
 * Added a function under the game menu to get the mission queue unstuck if it gets stuck
+* Game will auto peek face down cards you play
 
 ### 1.0.1.x